tests/requirements-test.txt
attrs==21.4.0
azure-ai-formrecognizer==3.1.0b3
azure-common==1.1.27
azure-core==1.21.1
azure-cosmosdb-nspkg==2.0.2
azure-cosmosdb-table==1.0.6
azure-functions==1.8.0
azure-nspkg==3.0.2
azure-storage-blob==12.9.0
blis==0.7.5
catalogue==2.0.6
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==7.1.2
cryptography==36.0.1
cymem==2.0.6
DateTime==4.3
ddt==1.4.4
edit-distance==1.0.4
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.0.0/en_core_web_sm-3.0.0-py3-none-any.whl
idna==3.3
iniconfig==1.1.1
isodate==0.6.1
jellyfish==0.9.0
Jinja2==3.0.3
Levenshtein==0.16.0
MarkupSafe==2.0.1
msrest==0.6.21
murmurhash==1.0.6
numpy==1.22.1
oauthlib==3.1.1
packaging==21.3
pathy==0.6.1
pluggy==1.0.0
preshed==3.0.6
py==1.11.0
pycparser==2.21
pydantic==1.8.2
pyparsing==3.0.6
pytest==6.2.5
python-dateutil==2.8.2
pytz==2021.3
rapidfuzz==1.8.3
requests==2.27.1
requests-oauthlib==1.3.0
six==1.16.0
smart-open==5.2.1
spacy==3.0.7
spacy-legacy==3.0.8
srsly==2.4.2
thinc==8.0.13
toml==0.10.2
tqdm==4.62.3
typer==0.3.2
typing_extensions==4.0.1
urllib3==1.26.8
wasabi==0.9.0
zope.interface==5.4.0